Refactor flask app structure
Move all functions to sub directory api
Seperate MVC structure of stories and users
Add database.py to access db related functions
FIx circular dependencies
