+++ fact-bounty-flask/api/__init__.py
@@ -1,5 +1,1 @@
-from flask import Blueprint
-
-api = Blueprint('api', __name__)
-
-from . import routes, errors
+"""Main application package."""
