+++ fact-bounty-flask/api/app.py
@@ -1,24 +1,17 @@
 from flask import Flask
-from flask_login import LoginManager
-from flask_pagedown import PageDown
-from flask_sqlalchemy import SQLAlchemy
 from elasticsearch import Elasticsearch
 from flask_cors import CORS
 
-from .config import config
-from . import commands
-
-db = SQLAlchemy()
-pagedown = PageDown()
-
-login_manager = LoginManager()
-login_manager.login_view = 'auth.login'
-
+from api import commands
+from api import user
+from api.config import config
+from api.extensions import db, pagedown, login_manager
 
 def create_app(config_name):
     # create and configure the app
     app = Flask(__name__)
     app.config.from_object(config[config_name])
+    config[config_name].init_app(app)
 
     register_extensions(app)
     register_blueprint(app)
@@ -37,8 +30,8 @@ def register_extensions(app):
 
 def register_blueprint(app):
     """Register Flask blueprints."""
-    from .api import api as api_blueprint
-    app.register_blueprint(api_blueprint, url_prefix='/api')
+    app.register_blueprint(user.views.blueprint, url_prefix='/api')
+    app.register_blueprint(stories.views.blueprint, url_prefix='/api')
     return None
 
 def register_commands(app):
