+++ fact-bounty-flask/api/errors.py
@@ -1,9 +1,5 @@
 from flask import jsonify
 
-from . import api
-from .exceptions import ValidationError
-
-
 def bad_request(message):
     response = jsonify({'error': 'bad request', 'message': message})
     response.status_code = 400
@@ -20,8 +16,3 @@ def forbidden(message):
     response = jsonify({'error': 'forbidden', 'message': message})
     response.status_code = 403
     return response
-
-
-@api.errorhandler(ValidationError)
-def validation_error(e):
-    return bad_request(e.args[0])
