--- fact-bounty-flask/api/exceptions.py
@@ -1,2 +1,0 @@
-class ValidationError(ValueError):
-    pass
