+++ fact-bounty-flask/api/stories/__init__.py
@@ -1,0 +1,1 @@
+from . import views
